Move validation to services
