+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryUserStorage.java
@@ -34,10 +34,7 @@ public class InMemoryUserStorage implements UserStorage {
 
     @Override
     public User create(User user) {
-
         log.info("Создание (post) записи для пользователя {}", user.getName());
-
-        validateUser(user);
 
         user.setId(++idCounter);
         users.put(user.getId(), user);
@@ -46,9 +43,7 @@ public class InMemoryUserStorage implements UserStorage {
 
     @Override
     public User update(User user) {
-
         log.info("Редактирование (put) записи для пользователя {}", user.getName());
-        validateUser(user);
 
         if (!users.containsKey(user.getId())) {
             throw new UserUnknownException("Пользователь с id = " + user.getId() + " не известен.");
@@ -59,26 +54,4 @@ public class InMemoryUserStorage implements UserStorage {
         return user;
     }
 
-    private void validateUser(User user) {
-
-        if (user.getEmail().isBlank() || !user.getEmail().contains("@")) {
-            log.info("User: Валидация не пройдена: email пуст или не содержит @");
-            throw new ValidationException("Электронная почта не может быть пустой и должна содержать @");
-        }
-
-        if (user.getLogin().isBlank() || user.getLogin().contains(" ")) {
-            log.info("User: Валидация не пройдена: логин пуст или содержит пробелы");
-            throw new ValidationException("Логин не может быть пустым и содержать пробелы.");
-        }
-
-        if (user.getName() == null || user.getName().isBlank()) {
-            log.info("User: Имя пользователя пустое: будет использоваться логин");
-            user.setName(user.getLogin());
-        }
-
-        if (user.getBirthday().isAfter(LocalDate.now())) {
-            log.info("User: Валидация не пройдена: дата рождения не может быть в будущем");
-            throw new ValidationException("Дата рождения не может быть в будущем");
-        }
-    }
 }
